{/.. is not considered to be in_re
See issue #1, but basically since `{\` does not trigger `in_re`, a `\\`
chain inside the regular expression literal will be interpreted as a
one line comment and removed completely (test case `{/\//}` which
results in a minified syntax error due to mismatched bracket when it's
turned into `{/\` (single line comment removed).

Not sure if this breaks anything but all tests pass.